refactor(stubs): split type stubs to match the three-module layout
Split the monolithic openscad stub into three packages mirroring
the runtime module hierarchy:

  * _openscad/__init__.pyi  - full canonical declarations, moved
                              verbatim from openscad/__init__.pyi.
  * openscad/__init__.pyi   - re-export shim that pulls every
                              public symbol from _openscad so
                              "from openscad import *" still type
                              checks.
  * pythonscad/__init__.pyi - same shim pattern, pulling from
                              openscad.

Update libraries/python/stubs/setup.py so the distributed stubs
package lists all three directories and bundles their py.typed
markers.

Made-with: Cursor
