+++ libraries/python/stubs/openscad/__init__.pyi
@@ -1,1815 +1,18 @@
-""" PythonSCAD Stub File for use in editors like Visual Studio Code """
-from typing import Union, Optional, overload
-
-PyOpenSCADs = Union["PyOpenSCAD", list["PyOpenSCAD"]]
-"""Type for functions that accept either a single OpenSCAD object or a list of objects."""
-
-Color = Union[str, list[float]]
-"""Color specification as either a color name string (e.g., "red") or RGB/RGBA values as [r, g, b] or [r, g, b, a]."""
-
-Vector2 = list[float]
-"""2D vector represented as [x, y] list."""
-
-Vector3 = list[float]
-"""3D vector represented as [x, y, z] list."""
-
-Matrix4x4 = list[list[float]]
-"""4x4 transformation matrix as a list of 4 lists of 4 floats."""
-
-
-
-class PyLibFive:
-    def __init__(self, x: float, y: float, z: float):
-        ...
-
-    def x(self) -> Self:
-        """Return X coordinate"""
-        ...
-
-    def y(self) -> Self:
-        """Return Y coordinate"""
-        ...
-
-    def z(self) -> Self:
-        """Return Z coordinate"""
-        ...
-
-    def sqrt(self, v) -> Self:
-        """Return Z coordinate"""
-        ...
-
-    def square(self, v) -> Self:
-        """Calculates the square of v"""
-        ...
-
-    def abs(self, v) -> Self:
-        """Calculates the absolute Value of v"""
-        ...
-
-    def max(self, a, b) -> Self:
-        """Calculates the maximum from a and b"""
-        ...
-
-    def min(self, a, b) -> Self:
-        """Calculates the minimum from a and b"""
-        ...
-
-    def sin(self, v) -> Self:
-        """Calculates the sine of v"""
-        ...
-
-    def cos(self, v) -> Self:
-        """Calculates the cosine of v"""
-        ...
-
-    def tan(self, v) -> Self:
-        """Calculates the tangents of v"""
-        ...
-
-    def asin(self, v) -> Self:
-        """Calculates the arc sine of v"""
-        ...
-
-    def acos(self, v) -> Self:
-        """Calculates the arc cosine of v"""
-        ...
-
-    def atan(self, v) -> Self:
-        """Calculates the arc tangent of v"""
-        ...
-
-    def exp(self, v) -> Self:
-        """Calculates the exponent of v"""
-        ...
-
-    def log(self, v) -> Self:
-        """Calculates the logarithm of v"""
-        ...
-
-    def pow(self, a, b) -> Self:
-        """Calculates the power from a over b"""
-        ...
-
-    def comp(self, a, b) -> Self:
-        """Calculates the libfive comparision of a and b"""
-        ...
-
-    def atan2(self, y, x) -> Self:
-        """Calculates aractangent from y/x"""
-        ...
-
-    def print(self) -> Self:
-        """Print the Tree for debugging"""
-        ...
-
-class PyOpenSCAD:
-    """Base class for OpenSCAD objects."""
-
-    origin: Matrix4x4
-    """4x4 transformation matrix representing the object's origin.
-    Initialized as identity matrix."""
-
-    def translate(self, v: Vector3) -> "PyOpenSCAD":
-        """Translate this object.
-
-        Args:
-            v: Translation vector [x, y, z].
-
-        Returns:
-            The transformed object. The original object is unaffected.
-        """
-        ...
-
-    def right(self, v: List[float]) -> Self:
-        """Moves an Object to the right"""
-        ...
-
-    def left(self, v: List[float]) -> Self:
-        """Moves an Object to the left"""
-        ...
-
-    def back(self, v: List[float]) -> Self:
-        """Moves Object backwards"""
-        ...
-
-    def front(self, v: List[float]) -> Self:
-        """Moves Object frontwards"""
-        ...
-
-    def up(self, v: List[float]) -> Self:
-        """Move Object upwards"""
-        ...
-
-    def down(self, v: List[float]) -> Self:
-        """Move Object downwards"""
-        ...
-
-    def rotx(self, v: List[float]) -> Self:
-        """Rotate Object around X Axis"""
-        ...
-
-    def roty(self, v: List[float]) -> Self:
-        """Rotate Object around Y Axis"""
-        ...
-
-    def rotz(self, v: List[float]) -> Self:
-        """Rotate Object around Z Axis"""
-        ...
-
-    def rotate(
-        self, a: Union[float, Vector3], v: Optional[Vector3] = None, ref: Optional[Vector3] = None
-    ) -> "PyOpenSCAD":
-        """Rotate this object.
-
-        Args:
-            a: Rotation angle (degrees) or rotation vector [x, y, z].
-            v: Optional rotation axis vector when a is a scalar angle.
-            ref: Optional specify center of rotation
-
-        Returns:
-            The transformed object. The original object is unaffected.
-        """
-        ...
-
-    def scale(self, v: Union[float, Vector3]) -> "PyOpenSCAD":
-        """Scale this object.
-
-        Args:
-            v: Scale factor (uniform) or scale vector [x, y, z].
-
-        Returns:
-            The transformed object. The original object is unaffected.
-        """
-        ...
-
-    def mirror(self, v: Vector3) -> "PyOpenSCAD":
-        """Mirror this object.
-
-        Args:
-            v: Mirror plane normal vector [x, y, z].
-
-        Returns:
-            The transformed object. The original object is unaffected.
-        """
-        ...
-
-    def multmatrix(self, m: Matrix4x4) -> "PyOpenSCAD":
-        """Apply matrix transformation to this object.
-
-        Args:
-            m: 4x4 transformation matrix as a list of 4 lists of 4 floats.
-
-        Returns:
-            The transformed object. The original object is unaffected.
-        """
-        ...
-
-    def divmatrix(self, m: Matrix4x4) -> "PyOpenSCAD":
-        """Apply inverse matrix transformation to this object.
-
-        Args:
-            m: 4x4 matrix as a list of 4 lists of 4 floats for inverse transformation.
-
-        Returns:
-            The transformed object. The original object is unaffected.
-        """
-        ...
-
-    def offset(
-        self,
-        r: Optional[float] = None,
-        delta: Optional[float] = None,
-        chamfer: Optional[bool] = None,
-        fn: Optional[float] = None,
-        fa: Optional[float] = None,
-        fs: Optional[float] = None,
-    ) -> "PyOpenSCAD":
-        """Offset this 2D object.
-
-        Args:
-            r: Offset radius (rounded corners).
-            delta: Offset distance (sharp corners).
-            chamfer: If True, creates chamfered corners.
-            fn: Number of fragments for curved parts.
-            fa: Minimum angle for each fragment.
-            fs: Minimum size for each fragment.
-
-        Returns:
-            The transformed object. The original object is unaffected.
-        """
-        ...
-
-    class RoofMethod(Enum):
-        TOP = "top"
-        LOFT = "loft"
-    def roof(
-        self,
-        method: str = RoofMethod.TOP.name,
-        convexity: int = 2,
-        fn: int = FN,
-        fa: float = FA,
-        fs: float = FS,
-    ) -> Self:
-        """Create Roof from an 2D Shape"""
-        ...
-
-    def pull(
-        self,
-        src: Self,
-        dst: Self
-    ) -> Self:
-        """Pull apart Object, basically between src and dst it creates a prisma with the x-section
-        src: anchor
-        dst: how much to pull
-        """
-        ...
-
-    def color(self, c: Color, alpha: float = 1.0) -> "PyOpenSCAD":
-        """Color this object.
-
-        Args:
-            c: Color specification - color name string or RGB/RGBA values.
-            alpha: Alpha (transparency) value between 0.0 and 1.0. Defaults to 1.0.
-
-        Returns:
-            A new object with the color set. The original object is unaffected.
-        """
-        ...
-
-    def output(self) -> None:
-        """sme as show"""
-        ...
-
-    def show(self) -> None:
-        """Mark this object for output/display."""
-        ...
-
-    def export(self, file: str) -> None:
-        """Export the result to a file
-        file: output file name, format is automatically detected from suffix
-        when obj is a dictionary, it allows 3mf export to export several paths
-        """
-        ...
-
-    def linear_extrude(
-        self,
-        height: Optional[float] = None,
-        convexity: int = 1,
-        center: Optional[bool] = None,
-        slices: int = 1,
-        segments: int = 0,
-        scale: Optional[Vector2] = None,
-        twist: Optional[float] = None,
-        fn: Optional[float] = None,
-        fa: Optional[float] = None,
-        fs: Optional[float] = None,
-    ) -> "PyOpenSCAD":
-        """Linear extrude this 2D object to 3D.
-
-        Args:
-            height: Extrusion height.
-            convexity: Convexity parameter for rendering. Defaults to 1.
-            center: If True, centers the extrusion.
-            slices: Number of slices for twist/scale. Defaults to 1.
-            segments: Number of segments. Defaults to 0.
-            scale: Scale factor for top vs bottom.
-            twist: Twist angle in degrees.
-            fn: Number of fragments for curved parts.
-            fa: Minimum angle for each fragment.
-            fs: Minimum size for each fragment.
-
-        Returns:
-            A new object representing the result of the extrusion. The original object is unaffected.
-        """
-        ...
-
-    def rotate_extrude(
-        self,
-        convexity: int = 1,
-        angle: float = 360.0,
-        fn: Optional[float] = None,
-        fa: Optional[float] = None,
-        fs: Optional[float] = None,
-    ) -> "PyOpenSCAD":
-        """Rotationally extrude this 2D object to 3D.
-
-        Args:
-            convexity: Convexity parameter for rendering. Defaults to 1.
-            angle: Rotation angle in degrees. Defaults to 360.0.
-            fn: Number of fragments for circle approximation.
-            fa: Minimum angle for each fragment.
-            fs: Minimum size for each fragment.
-
-        Returns:
-            A new object representing the result of the extrusion. The original object is unaffected.
-        """
-        ...
-
-    def path_extrude(
-        self,
-        path: List[float],
-        xdir: List[float],
-        convexity: int = 2,
-        origin: List[float] = [0, 0, 0],
-        scale: float = 1,
-        twist: float = 0,
-        closed: bool = False,
-        fn: int = FN,
-        fa: float = FA,
-        fs: float = FS,
-    ) -> Self:
-        """Path_extrude an 2D Object
-        xdir: initial vector of x axis
-        """
-        ...
-
-    def resize(self, newsize: Vector3, convexity: int = 2) -> "PyOpenSCAD":
-        """Modifies the size of the object to match the given x,y, and z sizes.
-
-        Args:
-            newsize: New size dimensions as [x, y, z].
-            convexity: Convexity parameter for rendering. Defaults to 2.
-
-        Returns:
-            The transformed object. The original object is unaffected.
-        """
-        ...
-
-    def highlight(self) -> Self:
-        """Highlights Object"""
-        ...
-
-    def background(self) -> Self:
-        """Puts Object into background"""
-        ...
-
-    def only(self) -> Self:
-        """Shows only this object"""
-        ...
-
-    def projection(
-        self, cut: Optional[bool] = None, convexity: int = 2
-    ) -> "PyOpenSCAD":
-        """Create a 2D projection from this 3D object.
-
-        Args:
-            cut: If True, creates a cross-section at z=0.
-            convexity: Convexity parameter for rendering. Defaults to 2.
-
-        Returns:
-            The projected 2D object.
-        """
-        ...
-
-    def mesh(
-        self, triangulate: Optional[bool] = None
-    ) -> Union[tuple[list[Vector3], list[list[int]]], list[list[Vector2]]]:
-        """Export mesh representation of this object.
-
-        Args:
-            triangulate: If True, triangulates the mesh.
-
-        Returns:
-            For 3D objects: A tuple of (vertices, faces) where:
-            - vertices: List of 3D vertex coordinates [[x, y, z], ...]
-            - faces: List of face definitions (lists of vertex indices)
-
-            For 2D objects: A list of outlines where:
-            - Each outline is a list of 2D vertex coordinates [[x, y], ...]
-        """
-        ...
-
-    def oversample(
-        self,
-        n: Optional[int] = None,
-        round: Optional[bool] = None
-    ) -> Self:
-        """Create artificial intermediate points into straight lines
-        n: factor of the oversampling
-        round: whether to round the oversampling
-        """
-        ...
-
-    def fillet(
-        self,
-        r: Optional[float] = None,
-        sel: Optional[Self] = None,
-        fn: Optional[int] = None
-    ) -> Self:
-        """Create nice roundings for sharp edges
-        r: radius of the fillet
-        sel: Object which overlaps the "selected" edges
-        fn: number of points for fillet x-section
-        """
-        ...
-
-    def align(
-        self, refmat: Matrix4x4, objmat: Optional[Matrix4x4] = None
-    ) -> "PyOpenSCAD":
-        """Align this object to a reference matrix.
-
-        Args:
-            refmat: Reference transformation matrix.
-            objmat: Optional object transformation matrix.
-
-        Returns:
-            A new object. The original object is unaffected.
-        """
-        ...
-
-
-    def render(self, convexity: int = 2) -> "PyOpenSCAD":
-        """Force rendering this object.
-
-        Args:
-            convexity: Convexity parameter for rendering. Defaults to 2.
-
-        Returns:
-            The object that will be forced to render. The original object is unaffected.
-        """
-        ...
-
-    def union(self, *others: PyOpenSCADs) -> "PyOpenSCAD":
-        """Create a union of this object with others.
-
-        Args:
-            *others: Other OpenSCAD objects to union with this one.
-
-        Returns:
-            A new object representing the union. The original object is unaffected.
-        """
-        ...
-
-    def difference(self, *others: PyOpenSCADs) -> "PyOpenSCAD":
-        """Create a difference by subtracting others from this object.
-
-        Args:
-            *others: Other OpenSCAD objects to subtract from this one.
-
-        Returns:
-            A new object representing the difference. The original object is unaffected.
-        """
-        ...
-
-    def intersection(self, *others: PyOpenSCADs) -> "PyOpenSCAD":
-        """Create an intersection of this object with others.
-
-        Args:
-            *others: Other OpenSCAD objects to intersect with this one.
-
-        Returns:
-            A new object representing the intersection. The original object is unaffected.
-        """
-        ...
-
-    def __getattr__(self, name): ...
-    def __setattr__(self, name, value): ...
-    def __getitem__(self, name): ...
-    def __setitem__(self, name, value): ...
-
-    # Operators:
-
-    def __or__(self, other: PyOpenSCADs) -> "PyOpenSCAD":
-        """Create a union of two objects"""
-        ...
-
-    def __and__(self, other: PyOpenSCADs) -> "PyOpenSCAD":
-        """Create an intersection of two objects"""
-        ...
-
-    @overload
-    def __sub__(self, other: PyOpenSCADs) -> "PyOpenSCAD":
-        """Create a difference of two objects"""
-        ...
-
-    def __add__(self, other: Vector3) -> "PyOpenSCAD":
-        """Create a new object by translating this object by a vector"""
-        ...
-
-    @overload
-    def __sub__(self, other: Vector3) -> "PyOpenSCAD":
-        """Create a new object by translating this object by the negative of a vector"""
-        ...
-
-    @overload
-    def __mul__(self, other: float) -> "PyOpenSCAD":
-        """Create a new object by scaling this object by a uniform factor in all directions"""
-        ...
-
-    @overload
-    def __mul__(self, other: Vector3) -> "PyOpenSCAD":
-        """Create a new object by scaling this object by a vector of factors in [x, y, z] directions"""
-        ...
-
-    @overload
-    def __xor__(self, other: PyOpenSCAD) -> "PyOpenSCAD":
-        """Create the hull from self and other"""
-        ...
-
-    @overload
-    def __xor__(self, other: list[Vector3]) -> "PyOpenSCAD":
-        """Create the explosion object from self and the vector list"""
-        ...
-
-    @overload
-    def __mod__(self, other: PyOpenSCAD) -> "PyOpenSCAD":
-        """Create a new object from the minkowski sum from self and other"""
-        ...
-
-    @overload
-    def __mod__(self, other: Vector3) -> "PyOpenSCAD":
-        """Rotate object by the rotations given in x/y/z of the Vector"""
-        ...
-
-def square(
-    dim: Optional[Union[float, list[float]]] = None, center: Optional[bool] = None
-) -> PyOpenSCAD:
-    """Create a square primitive.
-
-    Args:
-        dim: Dimensions of the square. Can be a single number for a square,
-             or a sequence of 2 numbers [width, height] for a rectangle.
-             If not specified, creates a unit square.
-        center: If True, centers the square at the origin. If False or None,
-                places one corner at the origin. Defaults to False.
-
-    Returns:
-        A 2D geometric object.
-    """
-    ...
-
-def circle(
-    r: Optional[float] = None,
-    d: Optional[float] = None,
-    fn: Optional[float] = None,
-    fa: Optional[float] = None,
-    fs: Optional[float] = None,
-) -> PyOpenSCAD:
-    """Create a circle primitive.
-
-    Args:
-        r: Radius of the circle. Must be positive. Cannot be used with d.
-             If neither r nor d is specified, defaults to 1.
-        d: Diameter of the circle. Must be positive. Cannot be used with r.
-        fn: Number of fragments for circle approximation.
-        fa: Minimum angle for each fragment.
-        fs: Minimum size for each fragment.
-
-    Returns:
-        A 2D geometric object.
-    """
-    ...
-
-def polygon(
-    points: Matrix4x4, paths: Optional[list[list[int]]] = None, convexity: int = 2
-) -> PyOpenSCAD:
-    """Create a polygon primitive.
-
-    Args:
-        points: List of 2D coordinates defining the polygon vertices.
-                Each point must be a list of exactly 2 numbers [x, y].
-                Must contain at least one point.
-        paths: Optional list of paths, where each path is a list of indices
-               into the points list. If specified, must contain at least one path.
-               Used to define holes or complex polygons.
-        convexity: Convexity parameter for rendering optimization. Must be >= 1.
-                   Defaults to 2.
-
-    Returns:
-        A 2D geometric object.
-    """
-    ...
-
-def text(
-    text: str,
-    size: float = 1.0,
-    font: Optional[str] = None,
-    spacing: float = 1.0,
-    direction: str = "ltr",
-    language: str = "en",
-    script: str = "latin",
-    halign: str = "left",
-    valign: str = "baseline",
-    fn: Optional[float] = None,
-    fa: Optional[float] = None,
-    fs: Optional[float] = None,
-) -> PyOpenSCAD:
-    """Create a text primitive.
-
-    Args:
-        text: The text string to render.
-        size: Font size. Defaults to 1.0.
-        font: Font name to use. If None, uses default font.
-        spacing: Spacing between characters. Defaults to 1.0.
-        direction: Text direction, either "ltr" (left-to-right) or "rtl". Defaults to "ltr".
-        language: Language code (e.g., "en", "de"). Defaults to "en".
-        script: Script type (e.g., "latin", "arabic"). Defaults to "latin".
-        halign: Horizontal alignment: "left", "center", or "right". Defaults to "left".
-        valign: Vertical alignment: "baseline", "top", "center", or "bottom". Defaults to "baseline".
-        fn: Number of fragments for curved parts.
-        fa: Minimum angle for each fragment.
-        fs: Minimum size for each fragment.
-
-    Returns:
-        A 2D geometric object.
-    """
-    ...
-
-def textmetrics(
-    text: str,
-    size: float = 1.0,
-    font: Optional[str] = None,
-    spacing: float = 1.0,
-    direction: str = "ltr",
-    language: str = "en",
-    script: str = "latin",
-    halign: str = "left",
-    valign: str = "baseline",
-) -> dict[str, Union[float, list[float]]]:
-    """Get text metrics for the given text parameters.
-
-    Args:
-        text: The text string to measure.
-        size: Font size. Defaults to 1.0.
-        font: Font name to use. If None, uses default font.
-        spacing: Spacing between characters. Defaults to 1.0.
-        direction: Text direction, either "ltr" or "rtl". Defaults to "ltr".
-        language: Language code (e.g., "en", "de"). Defaults to "en".
-        script: Script type (e.g., "latin", "arabic"). Defaults to "latin".
-        halign: Horizontal alignment: "left", "center", or "right". Defaults to "left".
-        valign: Vertical alignment: "baseline", "top", "center", or "bottom". Defaults to "baseline".
-
-    Returns:
-        A dictionary containing text metrics with keys:
-        - "ascent": Font ascent value
-        - "descent": Font descent value
-        - "offset": [x_offset, y_offset] list
-        - "advance": [advance_x, advance_y] list
-        - "position": [bbox_x, bbox_y] list
-        - "size": [bbox_width, bbox_height] list
-    """
-    ...
-
-def cube(
-    size: Optional[Union[float, Vector3]] = None, center: Optional[bool] = None
-) -> PyOpenSCAD:
-    """Create a cube primitive.
-
-    Args:
-        size: Dimensions of the cube. Can be a single number for a cube,
-              or a sequence of 3 numbers [x, y, z] for a rectangular box.
-              If not specified, creates a unit cube [1, 1, 1].
-        center: If True, centers the cube at the origin. If False or None,
-                places one corner at the origin. Defaults to False.
-
-    Returns:
-        A 3D geometric object.
-    """
-    ...
-
-def cylinder(
-    h: Optional[float] = None,
-    r1: Optional[float] = None,
-    r2: Optional[float] = None,
-    center: Optional[bool] = None,
-    r: Optional[float] = None,
-    d: Optional[float] = None,
-    d1: Optional[float] = None,
-    d2: Optional[float] = None,
-    fn: Optional[float] = None,
-    fa: Optional[float] = None,
-    fs: Optional[float] = None,
-) -> PyOpenSCAD:
-    """Create a cylinder primitive.
-
-    Args:
-        h: Height of the cylinder. Must be positive.
-        r1: Radius at bottom. Must be non-negative.
-        r2: Radius at top. Must be non-negative.
-        center: If True, centers the cylinder at the origin.
-        r: Uniform radius for both ends.
-        d: Uniform diameter for both ends.
-        d1: Diameter at bottom.
-        d2: Diameter at top.
-        fn: Number of fragments for circle approximation.
-        fa: Minimum angle for each fragment.
-        fs: Minimum size for each fragment.
-
-    Returns:
-        A 3D geometric object.
-    """
-    ...
-
-def sphere(
-    r: Optional[float] = None,
-    d: Optional[float] = None,
-    fn: Optional[float] = None,
-    fa: Optional[float] = None,
-    fs: Optional[float] = None,
-) -> PyOpenSCAD:
-    """Create a sphere primitive.
-
-    Args:
-        r: Radius of the sphere. Must be positive. Cannot be used with d.
-        d: Diameter of the sphere. Must be positive. Cannot be used with r.
-        fn: Number of fragments for sphere approximation.
-        fa: Minimum angle for each fragment.
-        fs: Minimum size for each fragment.
-
-    Returns:
-        A 3D geometric object.
-    """
-    ...
-
-def polyhedron(
-    points: Matrix4x4,
-    faces: list[list[int]],
-    convexity: int = 2,
-    triangles: Optional[list[list[int]]] = None,
-) -> PyOpenSCAD:
-    """Create a polyhedron primitive.
-
-    Args:
-        points: List of 3D coordinates defining the polyhedron vertices.
-                Each point must be a list of exactly 3 numbers [x, y, z].
-                Must contain at least one point.
-        faces: List of face definitions, where each face is a list of indices
-               into the points list.
-        convexity: Convexity parameter for rendering optimization. Defaults to 2.
-        triangles: Optional backwards compatibility parameter for triangular faces.
-
-    Returns:
-        A 3D geometric object.
-    """
-    ...
-
-def frep(exp: PyLibFive, min: List[float], max: List[float], res: int) -> PyOpenSCAD:
-    """Create F-Rep (libfive)
-    exp : an SDF epression composed from SDF variables and operators, see tutorial
-    """
-    ...
-
-def ifrep(obj: PyOpenSCAD) -> PyLibFive:
-    """Create Inverse F-Rep(experimental)"""
-    ...
-
-@overload
-def translate(obj: PyOpenSCADs, v: Vector3) -> PyOpenSCAD:
-    """Translate an object or list of objects.
-
-    Args:
-        obj: Object or list of objects to translate.
-        v: Translation vector [x, y, z].
-
-    Returns:
-        The transformed object. The original object is unaffected.
-    """
-    ...
-
-@overload
-def translate(matrix: Matrix4x4, v: Vector3) -> Matrix4x4:
-    """Apply translation to a 4x4 transformation matrix.
-
-    Args:
-        matrix: 4x4 transformation matrix to translate.
-        v: Translation vector [x, y, z].
-
-    Returns:
-        The transformed matrix with translation applied.
-    """
-    ...
-
-
-def right(obj: PyOpenSCADType, v: List[float]) -> PyOpenSCAD:
-    """Moves an Object to the right"""
-    ...
-
-def left(obj: PyOpenSCADType, v: List[float]) -> PyOpenSCAD:
-    """Moves an Object to the left"""
-    ...
-
-def back(obj: PyOpenSCADType, v: List[float]) -> PyOpenSCAD:
-    """Moves Object backwards"""
-    ...
-
-def front(obj: PyOpenSCADType, v: List[float]) -> PyOpenSCAD:
-    """Moves Object frontwards"""
-    ...
-
-def up(obj: PyOpenSCADType, v: List[float]) -> PyOpenSCAD:
-    """Move Object upwards"""
-    ...
-
-def down(obj: PyOpenSCADType, v: List[float]) -> PyOpenSCAD:
-    """Move Object downwards"""
-    ...
-
-@overload
-def rotate(
-    obj: PyOpenSCADs, a: Union[float, Vector3], v: Optional[Vector3] = None
-) -> PyOpenSCAD:
-    """Rotate an object or list of objects.
-
-    Args:
-        obj: Object or list of objects to rotate.
-        a: Rotation angle (degrees) or rotation vector [x, y, z].
-        v: Optional rotation axis vector when a is a scalar angle.
-
-    Returns:
-        The transformed object. The original object is unaffected.
-    """
-    ...
-
-@overload
-def rotate(
-    matrix: Matrix4x4, a: Union[float, Vector3], v: Optional[Vector3] = None
-) -> Matrix4x4:
-    """Apply rotation to a 4x4 transformation matrix.
-
-    Args:
-        matrix: 4x4 transformation matrix to rotate.
-        a: Rotation angle (degrees) or rotation vector [x, y, z].
-        v: Optional rotation axis vector when a is a scalar angle.
-
-    Returns:
-        The transformed matrix with rotation applied.
-    """
-    ...
-
-def rotx(obj: PyOpenSCADType, v: List[float]) -> PyOpenSCAD:
-    """Rotate Object around X Axis"""
-    ...
-
-def roty(obj: PyOpenSCADType, v: List[float]) -> PyOpenSCAD:
-    """Rotate Object around Y Axis"""
-    ...
-
-def rotz(obj: PyOpenSCADType, v: List[float]) -> PyOpenSCAD:
-    """Rotate Object around Z Axis"""
-    ...
-
-@overload
-def scale(obj: PyOpenSCADs, v: Union[float, Vector3]) -> PyOpenSCAD:
-    """Scale an object or list of objects.
-
-    Args:
-        obj: Object or list of objects to scale.
-        v: Scale factor (uniform) or scale vector [x, y, z].
-
-    Returns:
-        The transformed object. The original object is unaffected.
-    """
-    ...
-
-@overload
-def scale(matrix: Matrix4x4, v: Union[float, Vector3]) -> Matrix4x4:
-    """Apply scaling to a 4x4 transformation matrix.
-
-    Args:
-        matrix: 4x4 transformation matrix to scale.
-        v: Scale factor (uniform) or scale vector [x, y, z].
-
-    Returns:
-        The transformed matrix with scaling applied.
-    """
-    ...
-
-@overload
-def mirror(obj: PyOpenSCADs, v: Vector3) -> PyOpenSCAD:
-    """Mirror an object or list of objects.
-
-    Args:
-        obj: Object or list of objects to mirror.
-        v: Mirror plane normal vector [x, y, z].
-
-    Returns:
-        The transformed object. The original object is unaffected.
-    """
-    ...
-
-@overload
-def mirror(matrix: Matrix4x4, v: Vector3) -> Matrix4x4:
-    """Apply mirroring to a 4x4 transformation matrix.
-
-    Args:
-        matrix: 4x4 transformation matrix to mirror.
-        v: Mirror plane normal vector [x, y, z].
-
-    Returns:
-        The transformed matrix with mirroring applied.
-    """
-    ...
-
-def multmatrix(obj: PyOpenSCADs, m: Matrix4x4) -> PyOpenSCAD:
-    """Apply matrix transformation to an object.
-
-    Args:
-        obj: Object to transform.
-        m: 4x4 transformation matrix as a list of 4 lists of 4 floats.
-
-    Returns:
-        The transformed object. The original object is unaffected.
-    """
-    ...
-
-def color(obj: PyOpenSCADs, c: Color, alpha: float = 1.0) -> PyOpenSCAD:
-    """Color an object.
-
-    Args:
-        obj: Object to color.
-        c: Color specification - color name string or RGB/RGBA values.
-        alpha: Alpha (transparency) value between 0.0 and 1.0. Defaults to 1.0.
-
-    Returns:
-        A new object with the color set. The original object is unaffected.
-    """
-    ...
-
-def union(*objects: PyOpenSCADs) -> PyOpenSCAD:
-    """Create a union of multiple objects.
-
-    Args:
-        *objects: Variable number of OpenSCAD objects or lists of objects to union.
-
-    Returns:
-        A new object representing the union. The original object is unaffected.
-    """
-    ...
-
-def difference(*objects: PyOpenSCADs) -> PyOpenSCAD:
-    """Create a difference of multiple objects.
-
-    Args:
-        *objects: Variable number of OpenSCAD objects or lists of objects. The first object
-                 has all subsequent objects subtracted from it.
-
-    Returns:
-        A new object representing the difference. The original object is unaffected.
-    """
-    ...
-
-def intersection(*objects: PyOpenSCADs) -> PyOpenSCAD:
-    """Create an intersection of multiple objects.
-
-    Args:
-        *objects: Variable number of OpenSCAD objects or lists of objects to intersect.
-
-    Returns:
-        A new object representing the intersection. The original object is unaffected.
-    """
-    ...
-
-def hull(*objects: PyOpenSCADs) -> PyOpenSCAD:
-    """Create a convex hull of multiple objects.
-
-    Args:
-        *objects: Variable number of OpenSCAD objects or lists of objects.
-
-    Returns:
-        A new object. The original object is unaffected.
-    """
-    ...
-
-def linear_extrude(
-    obj: PyOpenSCADs,
-    height: Optional[float] = None,
-    convexity: int = 1,
-    center: Optional[bool] = None,
-    slices: int = 1,
-    segments: int = 0,
-    scale: Optional[Vector2] = None,
-    twist: Optional[float] = None,
-    fn: Optional[float] = None,
-    fa: Optional[float] = None,
-    fs: Optional[float] = None,
-) -> PyOpenSCAD:
-    """Linear extrude a 2D object to 3D.
-
-    Args:
-        obj: 2D object to extrude.
-        height: Extrusion height.
-        convexity: Convexity parameter for rendering. Defaults to 1.
-        center: If True, centers the extrusion.
-        slices: Number of slices for twist/scale. Defaults to 1.
-        segments: Number of segments. Defaults to 0.
-        scale: Scale factor for top vs bottom.
-        twist: Twist angle in degrees.
-        fn: Number of fragments for curved parts.
-        fa: Minimum angle for each fragment.
-        fs: Minimum size for each fragment.
-
-    Returns:
-        The linearly extruded 3D object. The original object is unaffected.
-    """
-    ...
-
-def rotate_extrude(
-    obj: PyOpenSCADs,
-    convexity: int = 1,
-    angle: float = 360.0,
-    fn: Optional[float] = None,
-    fa: Optional[float] = None,
-    fs: Optional[float] = None,
-) -> PyOpenSCAD:
-    """Rotationally extrude a 2D object to 3D.
-
-    Args:
-        obj: 2D object to extrude.
-        convexity: Convexity parameter for rendering. Defaults to 1.
-        angle: Rotation angle in degrees. Defaults to 360.0.
-        fn: Number of fragments for circle approximation.
-        fa: Minimum angle for each fragment.
-        fs: Minimum size for each fragment.
-
-    Returns:
-        The rotationally extruded 3D object. The original object is unaffected.
-    """
-    ...
-
-def path_extrude(
-    obj: PyOpenSCADType,
-    path: List[float],
-    xdir: List[float],
-    convexity: int,
-    origin: List[float],
-    scale: float,
-    twist: float,
-    closed: bool,
-    fn: int,
-    fa: float,
-    fs: float,
-) -> PyOpenSCAD:
-    """Path_extrude an 2D Object
-    xdir: intial vector of x axis
-    """
-    ...
-
-def offset(
-    obj: PyOpenSCADs,
-    r: Optional[float] = None,
-    delta: Optional[float] = None,
-    chamfer: Optional[bool] = None,
-    fn: Optional[float] = None,
-    fa: Optional[float] = None,
-    fs: Optional[float] = None,
-) -> PyOpenSCAD:
-    """Offset a 2D object.
-
-    Args:
-        obj: 2D object to offset.
-        r: Offset radius (rounded corners).
-        delta: Offset distance (sharp corners).
-        chamfer: If True, creates chamfered corners.
-        fn: Number of fragments for curved parts.
-        fa: Minimum angle for each fragment.
-        fs: Minimum size for each fragment.
-
-    Returns:
-        The offset 2D object. The original object is unaffected.
-    """
-    ...
-
-def minkowski(*objects: PyOpenSCADs, convexity: int = 2) -> PyOpenSCAD:
-    """Create a Minkowski sum of objects.
-
-    Args:
-        *objects: Objects or lists of objects to compute Minkowski sum of.
-        convexity: Convexity parameter for rendering. Defaults to 2.
-
-    Returns:
-        A new object representing the Minkowski sum. The original object is unaffected.
-    """
-    ...
-
-def projection(
-    obj: PyOpenSCADs, cut: Optional[bool] = None, convexity: int = 2
-) -> PyOpenSCAD:
-    """Create a 2D projection from a 3D object.
-
-    Args:
-        obj: 3D object to project.
-        cut: If True, creates a cross-section at z=0.
-        convexity: Convexity parameter for rendering. Defaults to 2.
-
-    Returns:
-        The projected 2D object.
-    """
-    ...
-
-def surface(
-    file: str,
-    center: Optional[bool] = None,
-    convexity: int = 2,
-    invert: Optional[bool] = None,
-) -> PyOpenSCAD:
-    """Create a surface from a heightmap file.
-
-    Args:
-        file: Path to the heightmap image file.
-        center: If True, centers the surface.
-        convexity: Convexity parameter for rendering. Defaults to 2.
-        invert: If True, inverts the heightmap.
-
-    Returns:
-        A 3d object generated from the imported height map.
-    """
-    ...
-
-def show(obj: PyOpenSCADs) -> None:
-    """Mark an object for output/display.
-
-    Args:
-        obj: Object to mark for output.
-    """
-    ...
-
-def render(obj: PyOpenSCADs, convexity: int = 2) -> PyOpenSCAD:
-    """Force rendering an object.
-
-    Args:
-        obj: Object to render.
-        convexity: Convexity parameter for rendering. Defaults to 2.
-
-    Returns:
-        The object that will be forced to render. The original object is unaffected.
-    """
-    ...
-
-def output(obj: PyOpenSCADType) -> None:
-    """same as show"""
-    ...
-def resize(
-    obj: PyOpenSCADs, newsize: Vector3, convexity: int = 2
-) -> PyOpenSCAD:
-    """Modifies the size of an object to match the given x,y, and z sizes.
-
-    Args:
-        obj: Object to resize.
-        newsize: New size dimensions as [x, y, z].
-        convexity: Convexity parameter for rendering. Defaults to 2.
-
-    Returns:
-        The resized object. The original object is unaffected.
-    """
-    ...
-
-def divmatrix(obj: PyOpenSCADs, m: Matrix4x4) -> PyOpenSCAD:
-    """Apply inverse matrix transformation to an object.
-
-    Args:
-        obj: Object to transform.
-        m: 4x4 matrix as a list of 4 lists of 4 floats for inverse transformation.
-
-    Returns:
-        The inverse transformed object.
-    """
-    ...
-
-def fill(*objects: PyOpenSCADs) -> PyOpenSCAD:
-    """Create a fill operation on objects.
-
-    Args:
-        *objects: Variable number of OpenSCAD objects or lists of objects to fill.
-
-    Returns:
-        A new object representing the fill operation result. The original object is unaffected.
-    """
-    ...
-
-
-def roof(
-    obj: PyOpenSCADType, method: str, convexity: int, fn: int, fa: float, fs: float
-) -> PyOpenSCAD:
-    """Create Roof from an 2D Shape"""
-    ...
-
-def pull(obj: PyOpenSCADType, src: List[float], dst: List[float]) -> PyOpenSCAD:
-    """Pull apart Object, basically between src and dst it creates a prisma with the x-section
-    src: anchor
-    dst: how much to pull
-    """
-    ...
-
-
-def export(obj: PyOpenSCADType, file: str) -> None:
-    """Export the result to a file
-    file:  output file name, format is automatically detected from suffix
-    when obj is a dictionary, it allows 3mf export to export several paths
-    """
-    ...
-
-def mesh(
-    obj: PyOpenSCADs, triangulate: Optional[bool] = None
-) -> Union[tuple[list[Vector3], list[list[int]]], list[list[Vector2]]]:
-    """Export mesh representation of an object.
-
-    Args:
-        obj: Object to convert to mesh.
-        triangulate: If True, triangulates the mesh.
-
-    Returns:
-        For 3D objects: A tuple of (vertices, faces) where:
-        - vertices: List of 3D vertex coordinates [[x, y, z], ...]
-        - faces: List of face definitions (lists of vertex indices)
-
-        For 2D objects: A list of outlines where:
-        - Each outline is a list of 2D vertex coordinates [[x, y], ...]
-    """
-
-def highlight(obj: PyOpenSCAD) -> PyOpenSCAD:
-    """Highlights Object"""
-    ...
-
-def background(obj: PyOpenSCAD) -> PyOpenSCAD:
-    """Puts Object into background"""
-    ...
-
-def only(obj: PyOpenSCAD) -> PyOpenSCAD:
-    """Shows only this object"""
-    ...
-
-
-
-def oversample(obj: PyOpenSCADType, n: int, round: bool) -> PyOpenSCAD:
-    """Create artificial intermediate points into straight lines
-    n: factor  of the oversampling
-    bool: whether to round the oversampling
-    """
-    ...
-
-def fillet(obj: PyOpenSCADType, r: float, sel: PyOpenSCADType, fn: int) -> PyOpenSCAD:
-    """Create nice roundings for sharp edges
-    r: radius of the fillet
-    sel: Object which overlaps the "selected" edges
-    fn: number of points for fillet x-section
-    """
-    ...
-
-def group(obj: PyOpenSCAD) -> PyOpenSCAD:
-    """Groups several Objects"""
-    ...
-
-
-def render(obj: PyOpenSCADs, convexity: int = 2) -> PyOpenSCAD:
-    """Force rendering an object.
-
-    Args:
-        obj: Object to render.
-        convexity: Convexity parameter for rendering. Defaults to 2.
-
-    Returns:
-        The object that will be forced to render. The original object is unaffected.
-    """
-    ...
-
-
-def rendervars(
-    vpd: Optional[float] = None,
-    vpf: Optional[float] = None,
-    vpr: Optional[Vector3] = None,
-    vpt: Optional[Vector3] = None,
-) -> None:
-    """Set camera/viewport parameters from Python code.
-
-    All parameters are optional. Only the ones provided will be updated;
-    the rest keep their current values.
-
-    Args:
-        vpd: Viewer distance (camera distance from the object).
-        vpf: Field of view angle in degrees.
-        vpr: Viewport rotation as [x, y, z] in degrees.
-        vpt: Viewport translation as [x, y, z] (the point the camera looks at).
-    """
-    ...
-
-def machineconfig(
-    config: None
-) -> None:
-    """Specify extended parameters to be used for gcode Export
-
-    Args:
-        config: Python Dict tree with settings
-
-    machineconfig({
-        "default":{
-            "property":{
-                "initCode":"G90", # Potential init code
-                "exitCode":"",    # Potential exit code
-                "feedAddX":-1,    # reduce feedrate away from laser tube
-                "feedAddY":-1,    # same
-            }
-        },
-        "stroke":{ # create mapping blue color to laserpower/laserfeed
-            "property":{
-                "color":0x0000ffff,
-                "feed":1000,
-                "power":50
-            }
-        }
-    })
-
-    """
-    ...
-
-def osimport(
-    file: str,
-    layer: str,
-    convexity: int,
-    origin: List[float],
-    scale: float,
-    width: float,
-    height: float,
-    filename: str,
-    center: bool,
-    dpi: float,
-    stroke: bool
-    id: int,
-) -> PyOpenSCAD:
-    """Imports Object from disc
-    Args:
-        stroke: defaults to true which turnes open SVG pathes to polygons. if set to false,
-                SVG pathes are converted to polylines instead.
-
-
-    """
-    ...
-
-def osuse(path: str) -> PyOpenSCAD:
-    """Import an OpenSCAD library, exposing its modules and functions.
-
-    Args:
-        path: Path to the OpenSCAD (.scad) file to import.
-
-    Returns:
-        An object with the library's modules and functions as attributes.
-        Modules can be called to create geometry, functions can be called
-        to compute and return values.
-
-    Example:
-        >>> lib = osuse("mylib.scad")
-        >>> # Call a module to create geometry
-        >>> obj = lib.my_module(r=10, h=20)
-        >>> # Call a function to compute a value
-        >>> width = lib.get_width()
-        >>> area = lib.calculate_area(width=5, height=10)
-    """
-    ...
-
-def version() -> List[float]:
-    """Outputs pythonscad Version"""
-    ...
-
-def version_num() -> List[float]:
-    """Outputs pythoncad Version"""
-    ...
-
-def add_parameter(name: str, default) -> None:
-    """Adds Parameter for use in Customizer"""
-    ...
-
-def scad(code: str) -> PyOpenSCAD:
-    """Evaluate Code in SCAD syntax"""
-    ...
-
-
-def align(
-    obj: PyOpenSCAD, refmat: Matrix4x4, objmat: Optional[Matrix4x4] -> None
-) -> PyOpenSCAD:
-    """Align an object to a reference matrix.
-
-    Args:
-        obj: Object to align.
-        refmat: Reference transformation matrix.
-        objmat: Optional object transformation matrix.
-
-    Returns:
-        A new object after alignment. The original object is unaffected.
-    """
-    ...
-
-def edge(size: Optional[Union[float, Vector2]] = None, center: Optional[bool] = None) -> PyOpenSCAD:
-    """Create a 1D edge primitive.
-
-    Args:
-        size: Length of the edge. Can be a single number or [length].
-        center: If True, centers the edge at the origin.
-
-    Returns:
-        A 1D geometric object.
-    """
-    ...
-
-def spline(points: list[Vector2], fn: Optional[int] = None, fa: Optional[float] = None, fs: Optional[float] = None) -> PyOpenSCAD:
-    """Create a spline curve from points.
-
-    Args:
-        points: List of 2D control points.
-        fn: Number of fragments for curve approximation.
-        fa: Minimum angle for each fragment.
-        fs: Minimum size for each fragment.
-
-    Returns:
-        A 2D geometric object representing the spline.
-    """
-    ...
-
-def wrap(
-    obj: PyOpenSCADs,
-    target: Optional[PyOpenSCAD] = None,
-    r: Optional[float] = None,
-    d: Optional[float] = None,
-    fn: Optional[float] = None,
-    fa: Optional[float] = None,
-    fs: Optional[float] = None,
-) -> PyOpenSCAD:
-    """Wrap an object around a cylinder.
-
-    Args:
-        obj: Object to wrap.
-        target: Optional target object to wrap around.
-        r: Radius of the cylinder.
-        d: Diameter of the cylinder.
-        fn: Number of fragments for circle approximation.
-        fa: Minimum angle for each fragment.
-        fs: Minimum size for each fragment.
-
-    Returns:
-        The wrapped object.
-    """
-    ...
-
-def separate(obj: PyOpenSCADs) -> list[PyOpenSCAD]:
-    """Split an object into separate parts.
-
-    Args:
-        obj: Object to split into parts.
-
-    Returns:
-        A list of separate objects, one for each connected component.
-    """
-    ...
-
-def find_face(obj: PyOpenSCADs, vec: Vector3) -> PyOpenSCAD:
-    """Find a face on an object facing a given direction.
-
-    Args:
-        obj: Object to search for faces.
-        vec: Direction vector to find face normal.
-
-    Returns:
-        Information about the found face.
-    """
-    ...
-
-def sitonto(
-    obj: PyOpenSCADs,
-    vecz: Vector3,
-    vecx: Optional[Vector3] = None,
-    vecy: Optional[Vector3] = None,
-) -> PyOpenSCAD:
-    """Position an object with specified orientation vectors.
-
-    Args:
-        obj: Object to position.
-        vecz: Z-axis direction vector.
-        vecx: Optional X-axis direction vector.
-        vecy: Optional Y-axis direction vector.
-
-    Returns:
-        The positioned object.
-    """
-    ...
-
-def skin(*objects: PyOpenSCADs, convexity: int = 2) -> PyOpenSCAD:
-    """Create a skin surface connecting multiple 2D cross-sections.
-
-    Args:
-        *objects: Multiple 2D cross-section objects to connect.
-        convexity: Convexity parameter for rendering. Defaults to 2.
-
-    Returns:
-        A 3D object skinned across the cross-sections.
-    """
-    ...
-
-def concat(*objects: PyOpenSCADs) -> PyOpenSCAD:
-    """Concatenate multiple objects into a single object.
-
-    Args:
-        *objects: Variable number of objects to concatenate.
-
-    Returns:
-        A single object containing all input objects.
-    """
-    ...
-
-def sheet(
-    func,
-    imin: float,
-    imax: float,
-    jmin: float,
-    jmax: float,
-    fs: Optional[float] = None,
-    iclose: Optional[bool] = None,
-    jclose: Optional[bool] = None,
-) -> PyOpenSCAD:
-    """Create a parametric surface sheet.
-
-    Args:
-        func: Function that takes (i, j) and returns a 3D point.
-        imin: Minimum value for first parameter.
-        imax: Maximum value for first parameter.
-        jmin: Minimum value for second parameter.
-        jmax: Maximum value for second parameter.
-        fs: Fragment size for tesselation.
-        iclose: If True, closes the surface in the i direction.
-        jclose: If True, closes the surface in the j direction.
-
-    Returns:
-        A 3D parametric surface object.
-    """
-    ...
-
-def bbox(obj: PyOpenSCADs) -> list[Vector3]:
-    """Calculate the bounding box of an object.
-
-    Args:
-        obj: Object to calculate bounding box for.
-
-    Returns:
-        A list of two 3D vectors: [[min_x, min_y, min_z], [max_x, max_y, max_z]].
-    """
-    ...
-
-def size(obj: PyOpenSCADs) -> Vector3:
-    """Get the size dimensions of an object's bounding box.
-
-    Args:
-        obj: Object to measure.
-
-    Returns:
-        A 3D vector [width, height, depth] representing the object's dimensions.
-    """
-    ...
-
-def position(obj: PyOpenSCADs) -> Vector3:
-    """Get the position (minimum coordinates) of an object's bounding box.
-
-    Args:
-        obj: Object to get position from.
-
-    Returns:
-        A 3D vector [min_x, min_y, min_z] representing the minimum corner.
-    """
-    ...
-
-def faces(obj: PyOpenSCADs, triangulate: Optional[bool] = None) -> list:
-    """Export a list of faces from an object.
-
-    Args:
-        obj: Object to extract faces from.
-        triangulate: If True, triangulates the faces.
-
-    Returns:
-        A list of face definitions.
-    """
-    ...
-
-def edges(obj: PyOpenSCADs) -> list:
-    """Export a list of edges from an object or face.
-
-    Args:
-        obj: Object or face to extract edges from.
-
-    Returns:
-        A list of edges.
-    """
-    ...
-
-def explode(obj: PyOpenSCADs, v: list[Vector3]) -> PyOpenSCAD:
-    """Explode a solid with a vector list.
-
-    Args:
-        obj: Object to explode.
-        v: List of vectors for explosion directions.
-
-    Returns:
-        The exploded object.
-    """
-    ...
-
-def debug(obj: PyOpenSCADs, faces: Optional[list] = None) -> PyOpenSCAD:
-    """Debug an object or specific faces.
-
-    Args:
-        obj: Object to debug.
-        faces: Optional list of specific faces to debug.
-
-    Returns:
-        Debug visualization of the object.
-    """
-    ...
-
-def repair(obj: PyOpenSCADs, color: Optional[Color] = None) -> PyOpenSCAD:
-    """Repair an object to make it watertight.
-
-    Args:
-        obj: Object to repair.
-        color: Optional color for repaired areas.
-
-    Returns:
-        The repaired object.
-    """
-    ...
-
-def osinclude(path: str) -> PyOpenSCAD:
-    """Include an OpenSCAD library (deprecated - use osuse instead).
-
-    Args:
-        path: Path to the OpenSCAD (.scad) file to include.
-
-    Returns:
-        An object with the library's modules and functions as attributes.
-
-    Note:
-        This function is deprecated. Use osuse() instead.
-    """
-    ...
-
-def model() -> PyOpenSCAD:
-    """Yield the current model.
-
-    Returns:
-        The current model object.
-    """
-    ...
-
-def modelpath() -> str:
-    """Get the absolute path to the current script.
-
-    Returns:
-        Absolute path to the script file.
-    """
-    ...
-
-def memberfunction(membername: str, memberfunc, docstring: str) -> None:
-    """Register an additional OpenSCAD member function.
-
-    Args:
-        membername: Name of the member function.
-        memberfunc: Python function to register.
-        docstring: Documentation string for the function.
-    """
-    ...
-
-def marked(value: Union[float, Vector3, PyOpenSCAD]) -> PyOpenSCAD:
-    """Create a marked value for special processing.
-
-    Args:
-        value: Value to mark (number, vector, or object).
-
-    Returns:
-        A marked value object.
-    """
-    ...
-
-def Sin(value: float) -> float:
-    """Calculate sine of an angle in degrees.
-
-    Args:
-        value: Angle in degrees.
-
-    Returns:
-        Sine of the angle.
-    """
-    ...
-
-def Cos(value: float) -> float:
-    """Calculate cosine of an angle in degrees.
-
-    Args:
-        value: Angle in degrees.
-
-    Returns:
-        Cosine of the angle.
-    """
-    ...
-
-def Tan(value: float) -> float:
-    """Calculate tangent of an angle in degrees.
-
-    Args:
-        value: Angle in degrees.
-
-    Returns:
-        Tangent of the angle.
-    """
-    ...
-
-def Asin(value: float) -> float:
-    """Calculate arcsine in degrees.
-
-    Args:
-        value: Value between -1 and 1.
-
-    Returns:
-        Arcsine in degrees.
-    """
-    ...
-
-def Acos(value: float) -> float:
-    """Calculate arccosine in degrees.
-
-    Args:
-        value: Value between -1 and 1.
-
-    Returns:
-        Arccosine in degrees.
-    """
-    ...
-
-def Atan(value: float) -> float:
-    """Calculate arctangent in degrees.
-
-    Args:
-        value: Value to calculate arctangent for.
-
-    Returns:
-        Arctangent in degrees.
-    """
-    ...
-
-def norm(vec: Vector3) -> float:
-    """Calculate the length/magnitude of a vector.
-
-    Args:
-        vec: 3D vector [x, y, z].
-
-    Returns:
-        The length of the vector.
-    """
-    ...
-
-def dot(vec1: Vector3, vec2: Vector3) -> float:
-    """Calculate the dot product of two vectors.
-
-    Args:
-        vec1: First 3D vector [x, y, z].
-        vec2: Second 3D vector [x, y, z].
-
-    Returns:
-        The dot product of the two vectors.
-    """
-    ...
-
-def cross(vec1: Vector3, vec2: Vector3) -> Vector3:
-    """Calculate the cross product of two vectors.
-
-    Args:
-        vec1: First 3D vector [x, y, z].
-        vec2: Second 3D vector [x, y, z].
-
-    Returns:
-        The cross product as a 3D vector.
-    """
-    ...
+"""Type stubs for the `openscad` package.
+
+The `openscad` package re-exports `_openscad` 1:1 (plus, in the future,
+any pure-Python additions or overrides that should match upstream
+OpenSCAD's API). The canonical stubs live in `_openscad`; this module
+just re-exports them so editors see the same names regardless of which
+package the user imports from.
+"""
+from _openscad import *  # noqa: F401,F403
+from _openscad import (  # noqa: F401
+    Color,
+    Matrix4x4,
+    PyLibFive,
+    PyOpenSCAD,
+    PyOpenSCADs,
+    Vector2,
+    Vector3,
+)
